Refactor RecordWriter and FieldWriterSet
